+++ R/tar_make_archive.R
@@ -29,6 +29,8 @@ tar_make_archive <- function(
   terminate_controller = TRUE,
   as_job = targets::tar_config_get("as_job")
 ) {
+  check_string(package, allow_empty = FALSE)
+  check_string(pipeline, allow_empty = FALSE)
   script <- tar_archive_script(
     package = package,
     pipeline = pipeline,