+++ R/tar_load_archive.R
@@ -28,6 +28,7 @@ tar_load_archive <- function(
     pipeline = pipeline,
     store = store
   )
+  check_archive_store_exists(store, package = package, pipeline = pipeline)
   meta <- meta %||%
     targets::tar_meta(
       store = store