+++ tests/testthat/test-tar_load_archive.R
@@ -13,3 +13,14 @@ targets::tar_test("tar_load_archive() loads a target into an environment", {
   )
   expect_s3_class(envir$model, "lm")
 })
+
+targets::tar_test("tar_load_archive() errors when the store is not built", {
+  expect_snapshot(
+    error = TRUE,
+    tar_load_archive(
+      model,
+      package = "tarchives",
+      pipeline = "not-built"
+    )
+  )
+})