+++ R/tar_make_archive.R
@@ -34,6 +34,7 @@ tar_make_archive <- function(
   script <- tar_archive_script(
     package = package,
     pipeline = pipeline,
+    envir = envir,
     script = script
   )
   store <- tar_archive_store(
@@ -44,6 +45,7 @@ tar_make_archive <- function(
   with_dir_archive(
     package = package,
     pipeline = pipeline,
+    envir = envir,
     targets::tar_make(
       names = {{ names }},
       shortcut = shortcut,