+++ R/tar_destroy_archive.R
@@ -10,6 +10,17 @@
 #'
 #' @inherit targets::tar_destroy return
 #'
+#' @examples
+#' \donttest{
+#' withr::with_envvar(
+#'   c(R_USER_CACHE_DIR = tempfile()),
+#'   {
+#'     tar_make_archive(package = "tarchives", pipeline = "example-model")
+#'     tar_destroy_archive(package = "tarchives", pipeline = "example-model", ask = FALSE)
+#'   }
+#' )
+#' }
+#'
 #' @export
 tar_destroy_archive <- function(
   package,