+++ R/tar_meta_archive.R
@@ -9,6 +9,17 @@
 #'
 #' @inherit targets::tar_meta return
 #'
+#' @examples
+#' \donttest{
+#' withr::with_envvar(
+#'   c(R_USER_CACHE_DIR = tempfile()),
+#'   {
+#'     tar_make_archive(package = "tarchives", pipeline = "example-model")
+#'     tar_meta_archive(package = "tarchives", pipeline = "example-model")
+#'   }
+#' )
+#' }
+#'
 #' @export
 tar_meta_archive <- function(
   package,