+++ R/tar_read_archive.R
@@ -6,6 +6,17 @@
 #'
 #' @inherit targets::tar_read return
 #'
+#' @examples
+#' \donttest{
+#' withr::with_envvar(
+#'   c(R_USER_CACHE_DIR = tempfile()),
+#'   {
+#'     tar_make_archive(package = "tarchives", pipeline = "example-model")
+#'     tar_read_archive(model, package = "tarchives", pipeline = "example-model")
+#'   }
+#' )
+#' }
+#'
 #' @export
 tar_read_archive <- function(
   name,